Refactor: separate handlerComponent from eventHandler.js
